Changed behaviour controlled by 'bFill' in DiffusionLength.R.
Now, for 'TRUE' (default) the unknown gradients at the bottom of the depth
vector are set to the last known value before, no longer to zero, which gives
a more physical diffusion length at the bottom. For 'FALSE', the gradients are
set to 'NA' which also results in 'NA' for the final diffusion length value.