Update of Herron-Langway code DensityHL.R.
Implementation of the steady-state Herron-Langway densification solution was
checked and coding style improved. The flag' JohnsenCorr' was introduced to
account for the Johnsen et al. (2000) correction of the rate constants for
central Greenland sites. The documentation was largely updated. In addition,
minor edits of DiffusionLength.R and Diffusivity.R documentations.